\name{talkstats-package}
\alias{talkstats-package}
\alias{talkstats}
\docType{package}
\title{
talkstats member stuff
}
\description{
talkstats member stuff
}
\details{
\tabular{ll}{
Package: \tab talkstats\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-06\cr
License: \tab GPL-2\cr
}
}
\author{
Tyler Rinker

Maintainer: Tyler Rinker <tyler.rinker@gmail.com>
}

